+++ src/googleApi/GoogleUpdateTask.ts
@@ -2,8 +2,6 @@ import type GoogleTasksPlugin from "src/GoogleTasksPlugin";
 
 import { createNotice } from "src/helper/NoticeHelper";
 import type { Task } from "src/helper/types";
-import { GoogleTaskView, VIEW_TYPE_GOOGLE_TASK } from "src/view/GoogleTaskView";
-
 import { getGoogleAuthToken } from "./GoogleAuth";
 
 export async function UpdateGoogleTask(
@@ -39,14 +37,7 @@ export async function UpdateGoogleTask(
         if (response.status == 200) {
             createNotice(plugin, "Task updated");
             await response.json();
-    
-            plugin.app.workspace
-                .getLeavesOfType(VIEW_TYPE_GOOGLE_TASK)
-                .forEach((leaf) => {
-                    if (leaf.view instanceof GoogleTaskView) {
-                        leaf.view.onOpen();
-                    }
-                });
+
             plugin.refreshScheduleViews();
         }